FEAT click on the button added
The button now creates a prompt and catch the answer, and the answer gives an alert with the number catched of the input or "Mal" if the answer is not the expected.
In next commits the message given will change and the alert with the number will be removed and the SketchPad will be drawn.